+++ bbconf/modules/bedfiles.py
@@ -5,22 +5,20 @@ from typing import Dict, List, Union
 
 import numpy as np
 from geniml.bbclient import BBClient
-
-# from geniml.io import RegionSet
 from geniml.search.backends import QdrantBackend
 from gtars.models import RegionSet as GRegionSet
 from pephubclient.exceptions import ResponseError
 from pydantic import BaseModel
+from qdrant_client import models
 from qdrant_client.http.models import PointStruct
-from qdrant_client.models import Distance, PointIdsList, VectorParams
-from sqlalchemy import and_, delete, func, or_, select, cast
+from qdrant_client.http.exceptions import UnexpectedResponse
+from qdrant_client.models import PointIdsList, QueryResponse
+from sqlalchemy import and_, cast, delete, func, or_, select
+from sqlalchemy.dialects import postgresql
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session, aliased
 from sqlalchemy.orm.attributes import flag_modified
-from sqlalchemy.dialects import postgresql
 from tqdm import tqdm
-from fastembed import TextEmbedding
-from qdrant_client import models
 
 from bbconf.config_parser.bedbaseconfig import BedBaseConfig
 from bbconf.const import DEFAULT_LICENSE, PKG_NAME, ZARR_TOKENIZED_FOLDER
@@ -85,13 +83,9 @@ class BedAgentBedFile:
         """
         self._sa_engine = config.db_engine.engine
         self._db_engine = config.db_engine
-        self._qdrant_engine = config.qdrant_engine
         self._boto3_client = config.boto3_client
-        self._config = config
+        self.config = config
         self.bb_agent = bbagent_obj
-
-        self._embedding_model = TextEmbedding(config.config.path.text2vec)
-        # self._embedding_model = TextEmbedding("BAAI/bge-large-en-v1.5")
 
     def get(self, identifier: str, full: bool = False) -> BedMetadataAll:
         """
@@ -121,7 +115,7 @@ class BedAgentBedFile:
                             FileModel(
                                 **result.__dict__,
                                 object_id=f"bed.{identifier}.{result.name}",
-                                access_methods=self._config.construct_access_method_list(
+                                access_methods=self.config.construct_access_method_list(
                                     result.path
                                 ),
                             ),
@@ -135,7 +129,7 @@ class BedAgentBedFile:
                                 FileModel(
                                     **result.__dict__,
                                     object_id=f"bed.{identifier}.{result.name}",
-                                    access_methods=self._config.construct_access_method_list(
+                                    access_methods=self.config.construct_access_method_list(
                                         result.path
                                     ),
                                 ),
@@ -171,10 +165,10 @@ class BedAgentBedFile:
         try:
             if full:
                 bed_metadata = BedPEPHubRestrict(
-                    **self._config.phc.sample.get(
-                        namespace=self._config.config.phc.namespace,
-                        name=self._config.config.phc.name,
-                        tag=self._config.config.phc.tag,
+                    **self.config.phc.sample.get(
+                        namespace=self.config.config.phc.namespace,
+                        name=self.config.config.phc.name,
+                        tag=self.config.config.phc.tag,
                         sample_name=identifier,
                     )
                 )
@@ -250,7 +244,7 @@ class BedAgentBedFile:
                         FileModel(
                             **result.__dict__,
                             object_id=f"bed.{identifier}.{result.name}",
-                            access_methods=self._config.construct_access_method_list(
+                            access_methods=self.config.construct_access_method_list(
                                 result.path
                             ),
                         ),
@@ -272,23 +266,30 @@ class BedAgentBedFile:
         if not self.exists(identifier):
             raise BEDFileNotFoundError(f"Bed file with id: {identifier} not found.")
         s = identifier
-        results = self._qdrant_engine.qd_client.query_points(
-            collection_name=self._config.config.qdrant.file_collection,
-            query="-".join([s[:8], s[8:12], s[12:16], s[16:20], s[20:]]),
-            limit=limit,
-            offset=offset,
-        )
-        result_list = []
-        for result in results.points:
-            result_id = result.id.replace("-", "")
-            result_list.append(
-                QdrantSearchResult(
-                    id=result_id,
-                    payload=result.payload,
-                    score=result.score,
-                    metadata=self.get(result_id, full=False),
-                )
+        try:
+            results = self.config.qdrant_file_backend.qd_client.query_points(
+                collection_name=self.config.config.qdrant.file_collection,
+                query="-".join([s[:8], s[8:12], s[12:16], s[16:20], s[20:]]),
+                limit=limit,
+                offset=offset,
             )
+            result_list = []
+            for result in results.points:
+                result_id = result.id.replace("-", "")
+                result_list.append(
+                    QdrantSearchResult(
+                        id=result_id,
+                        payload=result.payload,
+                        score=result.score,
+                        metadata=self.get(result_id, full=False),
+                    )
+                )
+        except UnexpectedResponse as err:
+            _LOGGER.error(
+                f"Qdrant request failed. Error: {err}. Returning empty result set."
+            )
+            result_list = []
+
         return BedListSearchResult(
             count=self.bb_agent.get_stats().bedfiles_number,
             limit=limit,
@@ -319,7 +320,7 @@ class BedAgentBedFile:
                         FileModel(
                             **result.__dict__,
                             object_id=f"bed.{identifier}.{result.name}",
-                            access_methods=self._config.construct_access_method_list(
+                            access_methods=self.config.construct_access_method_list(
                                 result.path
                             ),
                         ),
@@ -334,10 +335,10 @@ class BedAgentBedFile:
         :return: project metadata
         """
         try:
-            bed_metadata = self._config.phc.sample.get(
-                namespace=self._config.config.phc.namespace,
-                name=self._config.config.phc.name,
-                tag=self._config.config.phc.tag,
+            bed_metadata = self.config.phc.sample.get(
+                namespace=self.config.config.phc.namespace,
+                name=self.config.config.phc.name,
+                tag=self.config.config.phc.tag,
                 sample_name=identifier,
             )
         except Exception as e:
@@ -390,8 +391,8 @@ class BedAgentBedFile:
         """
         if not self.exists(identifier):
             raise BEDFileNotFoundError(f"Bed file with id: {identifier} not found.")
-        result = self._qdrant_engine.qd_client.retrieve(
-            collection_name=self._config.config.qdrant.file_collection,
+        result = self.config.qdrant_file_backend.qd_client.retrieve(
+            collection_name=self.config.config.qdrant.file_collection,
             ids=[identifier],
             with_vectors=True,
             with_payload=True,
@@ -629,12 +630,12 @@ class BedAgentBedFile:
         # Upload files to s3
         if upload_s3:
             if files:
-                files = self._config.upload_files_s3(
+                files = self.config.upload_files_s3(
                     identifier, files=files, base_path=local_path, type="files"
                 )
 
             if plots:
-                plots = self._config.upload_files_s3(
+                plots = self.config.upload_files_s3(
                     identifier, files=plots, base_path=local_path, type="plots"
                 )
         with Session(self._sa_engine) as session:
@@ -935,7 +936,7 @@ class BedAgentBedFile:
 
         _LOGGER.info("Updating bed file plots..")
         if plots:
-            plots = self._config.upload_files_s3(
+            plots = self.config.upload_files_s3(
                 bed_object.id, files=plots, base_path=local_path, type="plots"
             )
         plots_dict = plots.model_dump(
@@ -983,7 +984,7 @@ class BedAgentBedFile:
 
         _LOGGER.info("Updating bed files..")
         if files:
-            files = self._config.upload_files_s3(
+            files = self.config.upload_files_s3(
                 bed_object.id, files=files, base_path=local_path, type="files"
             )
 
@@ -1105,16 +1106,16 @@ class BedAgentBedFile:
             self.delete_pephub_sample(identifier)
         if delete_qdrant:
             self.delete_qdrant_point(identifier)
-        self._config.delete_files_s3(files)
+        self.config.delete_files_s3(files)
 
     def upload_pephub(self, identifier: str, metadata: dict, overwrite: bool = False):
         if not metadata:
             _LOGGER.warning("No metadata provided. Skipping pephub upload..")
             return False
-        self._config.phc.sample.create(
-            namespace=self._config.config.phc.namespace,
-            name=self._config.config.phc.name,
-            tag=self._config.config.phc.tag,
+        self.config.phc.sample.create(
+            namespace=self.config.config.phc.namespace,
+            name=self.config.config.phc.name,
+            tag=self.config.config.phc.tag,
             sample_name=identifier,
             sample_dict=metadata,
             overwrite=overwrite,
@@ -1127,10 +1128,10 @@ class BedAgentBedFile:
             if not metadata:
                 _LOGGER.warning("No metadata provided. Skipping pephub upload..")
                 return None
-            self._config.phc.sample.update(
-                namespace=self._config.config.phc.namespace,
-                name=self._config.config.phc.name,
-                tag=self._config.config.phc.tag,
+            self.config.phc.sample.update(
+                namespace=self.config.config.phc.namespace,
+                name=self.config.config.phc.name,
+                tag=self.config.config.phc.tag,
                 sample_name=identifier,
                 sample_dict=metadata,
             )
@@ -1144,10 +1145,10 @@ class BedAgentBedFile:
         :param identifier: bed file identifier
         """
         try:
-            self._config.phc.sample.remove(
-                namespace=self._config.config.phc.namespace,
-                name=self._config.config.phc.name,
-                tag=self._config.config.phc.tag,
+            self.config.phc.sample.remove(
+                namespace=self.config.config.phc.namespace,
+                name=self.config.config.phc.name,
+                tag=self.config.config.phc.tag,
                 sample_name=identifier,
             )
         except ResponseError as e:
@@ -1172,12 +1173,12 @@ class BedAgentBedFile:
 
         _LOGGER.debug(f"Adding bed file to qdrant. bed_id: {bed_id}")
 
-        if not isinstance(self._qdrant_engine, QdrantBackend):
+        if not isinstance(self.config.qdrant_file_backend, QdrantBackend):
             raise QdrantInstanceNotInitializedError("Could not upload file.")
 
         bed_embedding = self._embed_file(bed_file)
 
-        self._qdrant_engine.load(
+        self.config.qdrant_file_backend.load(
             ids=[bed_id],
             vectors=bed_embedding,
             payloads=[{**payload}],
@@ -1193,11 +1194,11 @@ class BedAgentBedFile:
 
         :return np array of embeddings
         """
-        if self._qdrant_engine is None:
+        if self.config.qdrant_file_backend is None:
             raise QdrantInstanceNotInitializedError
-        if not self._config.r2v:
+        if not self.config.r2v_encoder:
             raise BedBaseConfError(
-                "Could not add add region to qdrant. Invalid type, or path. "
+                "Could not add region to qdrant. Invalid type, or path. "
             )
 
         if isinstance(bed_file, str):
@@ -1210,9 +1211,9 @@ class BedAgentBedFile:
             bed_region_set = bed_file
         else:
             raise BedBaseConfError(
-                "Could not add add region to qdrant. Invalid type, or path. "
+                "Could not add region to qdrant. Invalid type, or path. "
             )
-        bed_embedding = np.mean(self._config.r2v.encode(bed_region_set), axis=0)
+        bed_embedding = np.mean(self.config.r2v_encoder.encode(bed_region_set), axis=0)
         vec_dim = bed_embedding.shape[0]
         return bed_embedding.reshape(1, vec_dim)
 
@@ -1223,11 +1224,11 @@ class BedAgentBedFile:
         :param bed_file: bed file path or region set
         """
 
-        if self._config._umap_model is None:
+        if self.config.umap_encoder is None:
             raise BedBaseConfError("UMAP model is not initialized.")
 
         bed_embedding = self._embed_file(bed_file)
-        bed_umap = self._config._umap_model.transform(bed_embedding)
+        bed_umap = self.config.umap_encoder.transform(bed_embedding)
         return bed_umap
 
     def text_to_bed_search(
@@ -1250,7 +1251,9 @@ class BedAgentBedFile:
         """
         _LOGGER.info(f"Looking for: {query}")
 
-        results = self._config.bivec.query_search(query, limit=limit, offset=offset)
+        results = self.config.bivec_search_interface.query_search(
+            query, limit=limit, offset=offset
+        )
         results_list = []
         for result in results:
             result_id = result["id"].replace("-", "")
@@ -1271,8 +1274,8 @@ class BedAgentBedFile:
             )
 
         if with_metadata:
-            count = self._config._qdrant_advanced_engine.get_collection(
-                collection_name=self._config.config.qdrant.file_collection
+            count = self.config.qdrant_client.get_collection(
+                collection_name=self.config.config.qdrant.file_collection
             ).points_count
         else:
             count = 0
@@ -1298,7 +1301,7 @@ class BedAgentBedFile:
 
         :return: BedListSetResults
         """
-        results = self._config.b2bsi.query_search(
+        results = self.config.b2b_search_interface.query_search(
             region_set, limit=limit, offset=offset
         )
         results_list = []
@@ -1501,8 +1504,8 @@ class BedAgentBedFile:
                         pbar.set_description(
                             "Uploading points to qdrant using batch..."
                         )
-                        operation_info = self._config.qdrant_engine.qd_client.upsert(
-                            collection_name=self._config.config.qdrant.file_collection,
+                        operation_info = self.config.qdrant_file_backend.qd_client.upsert(
+                            collection_name=self.config.config.qdrant.file_collection,
                             points=points_list,
                         )
                         pbar.write("Uploaded batch to qdrant.")
@@ -1515,8 +1518,8 @@ class BedAgentBedFile:
                     pbar.update(1)
 
             _LOGGER.info("Uploading points to qdrant using batches...")
-            operation_info = self._config.qdrant_engine.qd_client.upsert(
-                collection_name=self._config.config.qdrant.file_collection,
+            operation_info = self.config.qdrant_file_backend.qd_client.upsert(
+                collection_name=self.config.config.qdrant.file_collection,
                 points=points_list,
             )
             assert operation_info.status == "completed"
@@ -1530,22 +1533,13 @@ class BedAgentBedFile:
         :return: None
         """
 
-        result = self._config.qdrant_engine.qd_client.delete(
-            collection_name=self._config.config.qdrant.file_collection,
+        result = self.config.qdrant_file_backend.qd_client.delete(
+            collection_name=self.config.config.qdrant.file_collection,
             points_selector=PointIdsList(
                 points=[identifier],
             ),
         )
-        return result
-
-    def create_qdrant_collection(self) -> bool:
-        """
-        Create qdrant collection for bed files.
-        """
-        return self._config.qdrant_engine.qd_client.create_collection(
-            collection_name=self._config.config.qdrant.file_collection,
-            vectors_config=VectorParams(size=100, distance=Distance.DOT),
-        )
+        return None
 
     def exists(self, identifier: str) -> bool:
         """
@@ -1641,13 +1635,12 @@ class BedAgentBedFile:
 
             if self.exist_tokenized(bed_id, universe_id):
                 if not overwrite:
-                    if not overwrite:
-                        raise TokenizeFileExistsError(
-                            "Tokenized file already exists in the database. "
-                            "Set overwrite to True to overwrite it."
-                        )
-                    else:
-                        self.delete_tokenized(bed_id, universe_id)
+                    raise TokenizeFileExistsError(
+                        "Tokenized file already exists in the database. "
+                        "Set overwrite to True to overwrite it."
+                    )
+                else:
+                    self.delete_tokenized(bed_id, universe_id)
 
             path = self._add_zarr_s3(
                 bed_id=bed_id,
@@ -1655,7 +1648,7 @@ class BedAgentBedFile:
                 tokenized_vector=token_vector,
                 overwrite=overwrite,
             )
-            path = os.path.join(f"s3://{self._config.config.s3.bucket}", path)
+            path = os.path.join(f"s3://{self.config.config.s3.bucket}", path)
             new_token = TokenizedBed(bed_id=bed_id, universe_id=universe_id, path=path)
 
             session.add(new_token)
@@ -1678,7 +1671,7 @@ class BedAgentBedFile:
 
         :return: zarr path
         """
-        univers_group = self._config.zarr_root.require_group(universe_id)
+        univers_group = self.config.zarr_root.require_group(universe_id)
 
         if not univers_group.get(bed_id):
             _LOGGER.info("Saving tokenized vector to s3")
@@ -1708,7 +1701,7 @@ class BedAgentBedFile:
 
         if not self.exist_tokenized(bed_id, universe_id):
             raise TokenizeFileNotExistError("Tokenized file not found in the database.")
-        univers_group = self._config.zarr_root.require_group(universe_id)
+        univers_group = self.config.zarr_root.require_group(universe_id)
 
         return TokenizedBedResponse(
             universe_id=universe_id,
@@ -1727,7 +1720,7 @@ class BedAgentBedFile:
         """
         if not self.exist_tokenized(bed_id, universe_id):
             raise TokenizeFileNotExistError("Tokenized file not found in the database.")
-        univers_group = self._config.zarr_root.require_group(universe_id)
+        univers_group = self.config.zarr_root.require_group(universe_id)
 
         del univers_group[bed_id]
 
@@ -1801,7 +1794,7 @@ class BedAgentBedFile:
         file_path = self._get_tokenized_path(bed_id, universe_id)
 
         return TokenizedPathResponse(
-            endpoint_url=self._config.config.s3.endpoint_url,
+            endpoint_url=self.config.config.s3.endpoint_url,
             file_path=file_path,
             bed_id=bed_id,
             universe_id=universe_id,
@@ -1829,7 +1822,9 @@ class BedAgentBedFile:
             t2_alias = aliased(Files)
 
             # Define the subquery
-            subquery = select(t2_alias).where(t2_alias.name == plot_name).subquery()
+            subquery = (
+                select(t2_alias).where(and_(t2_alias.name == plot_name)).subquery()
+            )
 
             query = (
                 select(Bed.id)
@@ -1884,7 +1879,9 @@ class BedAgentBedFile:
             t2_alias = aliased(Files)
 
             # Define the subquery
-            subquery = select(t2_alias).where(t2_alias.name == "bigbed_file").subquery()
+            subquery = (
+                select(t2_alias).where(and_(t2_alias.name == "bigbed_file")).subquery()
+            )
 
             query = (
                 select(Bed.id)
@@ -2006,7 +2003,103 @@ class BedAgentBedFile:
 
             session.commit()
 
-    def reindex_semantic_search(self, batch: int = 1000, purge: bool = False) -> None:
+    # def reindex_semantic_search(self, batch: int = 1000, purge: bool = False) -> None:
+    #     """
+    #     Reindex all bed files for semantic database
+    #
+    #     :param batch: number of files to upload in one batch
+    #     :param purge: resets indexed in database for all files to False
+    #
+    #     :return: None
+    #     """
+    #
+    #     # Add column that will indicate if this file is indexed or not
+    #     statement = (
+    #         select(Bed)
+    #         .join(BedMetadata, Bed.id == BedMetadata.id)
+    #         .where(Bed.indexed == False)
+    #         .limit(150000)
+    #     )
+    #
+    #     with Session(self._sa_engine) as session:
+    #
+    #         if purge:
+    #             _LOGGER.info("Purging indexed files in the database ...")
+    #             session.query(Bed).update({Bed.indexed: False})
+    #             session.commit()
+    #             _LOGGER.info("Purged indexed files in the database successfully!")
+    #
+    #         _LOGGER.info("Fetching data from the database ...")
+    #         results = session.scalars(statement)
+    #
+    #         _LOGGER.info("Fetch data successfully!")
+    #
+    #         points = []
+    #         results = [result for result in results]
+    #
+    #         with tqdm(total=len(results), position=0, leave=True) as pbar:
+    #             processed_number = 0
+    #             for result in results:
+    #                 text = (
+    #                     f"biosample is {result.annotations.cell_line} / {result.annotations.cell_type} / "
+    #                     f"{result.annotations.tissue} with target {result.annotations.target} "
+    #                     f"assay {result.annotations.assay}."
+    #                     f"File name {result.name} with summary {result.description}"
+    #                 )
+    #
+    #                 embeddings_list = list(self.config.dense_encoder.embed(text))
+    #                 # result_list.append(
+    #                 data = VectorMetadata(
+    #                     id=result.id,
+    #                     name=result.name,
+    #                     description=result.description,
+    #                     genome_alias=result.genome_alias,
+    #                     genome_digest=result.genome_digest,
+    #                     cell_line=result.annotations.cell_line,
+    #                     cell_type=result.annotations.cell_type,
+    #                     tissue=result.annotations.tissue,
+    #                     target=result.annotations.target,
+    #                     treatment=result.annotations.treatment,
+    #                     assay=result.annotations.assay,
+    #                     species_name=result.annotations.species_name,
+    #                 )
+    #
+    #                 points.append(
+    #                     PointStruct(
+    #                         id=result.id,
+    #                         vector=list(embeddings_list[0]),
+    #                         payload=data.model_dump(),
+    #                     )
+    #                 )
+    #                 processed_number += 1
+    #                 result.indexed = True
+    #
+    #                 if processed_number % batch == 0:
+    #                     pbar.set_description(
+    #                         "Uploading points to qdrant using batch..."
+    #                     )
+    #                     operation_info = self.config._qdrant_advanced_engine.upsert(
+    #                         collection_name=self.config.config.qdrant.hybrid_collection,
+    #                         points=points,
+    #                     )
+    #                     session.commit()
+    #                     pbar.write("Uploaded batch to qdrant.")
+    #                     points = []
+    #                     assert operation_info.status == "completed"
+    #
+    #                 pbar.write(f"File: {result.id} successfully indexed.")
+    #                 pbar.update(1)
+    #
+    #         operation_info = self.config._qdrant_advanced_engine.upsert(
+    #             collection_name=self.config.config.qdrant.hybrid_collection,
+    #             points=points,
+    #         )
+    #         assert operation_info.status == "completed"
+    #         session.commit()
+    #
+    #     return None
+
+    def reindex_hybrid_search(self, batch: int = 1000, purge: bool = False) -> None:
         """
         Reindex all bed files for semantic database
 
@@ -2021,7 +2114,7 @@ class BedAgentBedFile:
             select(Bed)
             .join(BedMetadata, Bed.id == BedMetadata.id)
             .where(Bed.indexed == False)
-            .limit(150000)
+            .limit(batch)
         )
 
         with Session(self._sa_engine) as session:
@@ -2050,7 +2143,26 @@ class BedAgentBedFile:
                         f"File name {result.name} with summary {result.description}"
                     )
 
-                    embeddings_list = list(self._embedding_model.embed(text))
+                    embeddings_list = list(self.config.dense_encoder.embed(text))
+
+                    if self.config.sparse_encoder:
+                        sparse_result = self.config.sparse_encoder.encode(
+                            text
+                        ).coalesce()
+
+                        sparse_embeddings = models.SparseVector(
+                            indices=sparse_result.indices().tolist()[0],
+                            values=sparse_result.values().tolist(),
+                        )
+
+                        point_vectors = {
+                            "dense": list(embeddings_list[0]),
+                            "sparse": sparse_embeddings,
+                        }
+                    else:
+                        point_vectors = {
+                            "dense": list(embeddings_list[0]),
+                        }
                     # result_list.append(
                     data = VectorMetadata(
                         id=result.id,
@@ -2070,7 +2182,7 @@ class BedAgentBedFile:
                     points.append(
                         PointStruct(
                             id=result.id,
-                            vector=list(embeddings_list[0]),
+                            vector=point_vectors,
                             payload=data.model_dump(),
                         )
                     )
@@ -2081,8 +2193,8 @@ class BedAgentBedFile:
                         pbar.set_description(
                             "Uploading points to qdrant using batch..."
                         )
-                        operation_info = self._config._qdrant_advanced_engine.upsert(
-                            collection_name=self._config.config.qdrant.search_collection,
+                        operation_info = self.config.qdrant_client.upsert(
+                            collection_name=self.config.config.qdrant.hybrid_collection,
                             points=points,
                         )
                         session.commit()
@@ -2093,16 +2205,105 @@ class BedAgentBedFile:
                     pbar.write(f"File: {result.id} successfully indexed.")
                     pbar.update(1)
 
-            operation_info = self._config._qdrant_advanced_engine.upsert(
-                collection_name=self._config.config.qdrant.search_collection,
-                points=points,
-            )
-            assert operation_info.status == "completed"
+            if points:
+                operation_info = self.config.qdrant_client.upsert(
+                    collection_name=self.config.config.qdrant.hybrid_collection,
+                    points=points,
+                )
+                assert operation_info.status == "completed"
             session.commit()
 
         return None
 
-    def semantic_search(
+    # def semantic_search(
+    #     self,
+    #     query: str = "liver",
+    #     genome_alias: str = "",
+    #     assay: str = "",
+    #     limit: int = 100,
+    #     offset: int = 0,
+    #     with_metadata: bool = True,
+    # ) -> BedListSearchResult:
+    #     """
+    #     Run semantic search for bed files using qdrant.
+    #     This is not bivec search, but usual qdrant search with embeddings.
+    #
+    #     :param query: text query to search for
+    #     :param genome_alias: genome alias to filter results
+    #     :param assay: filter by assay type
+    #     :param limit: number of results to return
+    #     :param offset: offset to start from
+    #     :param with_metadata: if True, metadata will be returned in the results. Default is True.
+    #
+    #     :return: list of bed file metadata
+    #     """
+    #
+    #     should_statement = []
+    #
+    #     if genome_alias:
+    #         should_statement.append(
+    #             models.FieldCondition(
+    #                 key="genome_alias",
+    #                 match=models.MatchValue(value=genome_alias),
+    #             )
+    #         )
+    #     if assay:
+    #         should_statement.append(
+    #             models.FieldCondition(
+    #                 key="assay",
+    #                 match=models.MatchValue(value=assay),
+    #             )
+    #         )
+    #
+    #     embeddings_list = list(self.config.dense_encoder.embed(query))[0]
+    #
+    #     results: QueryResponse = self.config.qdrant_client.query_points(
+    #         collection_name=self.config.config.qdrant.hybrid_collection,
+    #         query=list(embeddings_list),
+    #         limit=limit,
+    #         offset=offset,
+    #         search_params=models.SearchParams(
+    #             exact=True,
+    #         ),
+    #         # query_filter=models.Filter(should=should_statement) if should_statement else None,
+    #         query_filter=(
+    #             models.Filter(must=should_statement) if should_statement else None
+    #         ),
+    #         with_payload=True,
+    #         with_vectors=True,
+    #     )
+    #
+    #     result_list = []
+    #     for result in results.points:
+    #         result_id = result.id.replace("-", "")
+    #
+    #         if with_metadata:
+    #             metadata = self.get(result_id, full=False)
+    #         else:
+    #             metadata = None
+    #
+    #         result_list.append(
+    #             QdrantSearchResult(
+    #                 id=result_id,
+    #                 payload=result.payload,
+    #                 score=result.score,
+    #                 metadata=metadata,
+    #             )
+    #         )
+    #
+    #     if with_metadata:
+    #         count = self.bb_agent.get_stats().bedfiles_number
+    #     else:
+    #         count = 0
+    #
+    #     return BedListSearchResult(
+    #         count=count,
+    #         limit=limit,
+    #         offset=offset,
+    #         results=result_list,
+    #     )
+
+    def hybrid_search(
         self,
         query: str = "liver",
         genome_alias: str = "",
@@ -2113,7 +2314,7 @@ class BedAgentBedFile:
     ) -> BedListSearchResult:
         """
         Run semantic search for bed files using qdrant.
-        This is not bivec search, but usual qdrant search with embeddings.
+        This is not bivec search, but usual qdrant search with sparse and dense embeddings.
 
         :param query: text query to search for
         :param genome_alias: genome alias to filter results
@@ -2125,43 +2326,67 @@ class BedAgentBedFile:
         :return: list of bed file metadata
         """
 
-        should_statement = []
+        must_statement = []
 
         if genome_alias:
-            should_statement.append(
+            must_statement.append(
                 models.FieldCondition(
                     key="genome_alias",
                     match=models.MatchValue(value=genome_alias),
                 )
             )
         if assay:
-            should_statement.append(
+            must_statement.append(
                 models.FieldCondition(
                     key="assay",
                     match=models.MatchValue(value=assay),
                 )
             )
 
-        embeddings_list = list(self._embedding_model.embed(query))[0]
+        dense_query = list(list(self.config.dense_encoder.embed(query))[0])
+        if self.config.sparse_encoder:
+            sparse_result = self.config.sparse_encoder.encode(query).coalesce()
+            sparse_embeddings = models.SparseVector(
+                indices=sparse_result.indices().tolist()[0],
+                values=sparse_result.values().tolist(),
+            )
 
-        results = self._config._qdrant_advanced_engine.search(
-            collection_name=self._config.config.qdrant.search_collection,
-            query_vector=list(embeddings_list),
+            hybrid_query = [
+                # Dense retrieval: semantic understanding
+                models.Prefetch(
+                    query=dense_query, using="dense", limit=limit + offset + 100
+                ),
+                # Sparse retrieval: exact technical term matching
+                models.Prefetch(
+                    query=sparse_embeddings, using="sparse", limit=limit + offset + 100
+                ),
+            ]
+        else:
+            hybrid_query = [
+                # Dense retrieval: semantic understanding
+                models.Prefetch(
+                    query=dense_query, using="dense", limit=limit + offset + 100
+                ),
+            ]
+
+        results = self.config.qdrant_client.query_points(
+            collection_name=self.config.config.qdrant.hybrid_collection,
             limit=limit,
             offset=offset,
+            prefetch=hybrid_query,
+            query=models.FusionQuery(fusion=models.Fusion.RRF),
+            with_payload=True,
+            with_vectors=True,
             search_params=models.SearchParams(
                 exact=True,
             ),
-            # query_filter=models.Filter(should=should_statement) if should_statement else None,
             query_filter=(
-                models.Filter(must=should_statement) if should_statement else None
+                models.Filter(must=must_statement) if must_statement else None
             ),
-            with_payload=True,
-            with_vectors=True,
         )
 
         result_list = []
-        for result in results:
+        for result in results.points:
             result_id = result.id.replace("-", "")
 
             if with_metadata: